Python/requirements.txt
cryptography==3.4.8
pygithub
boto3